refactor: Repository Layer 객체들을 domain 패키지로 이동
