+++ src/main/java/kitchenpos/domain/MenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.dao;
+package kitchenpos.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
